+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change log
 
 ## Release 0.7.0
+* New: Allow copying items between tabs. ([#318](https://github.com/ksnip/kImageAnnotator/issues/318))
 
 ## Release 0.6.1
 * Fixed: Fix for unnecessary scrollbars when a screenshot has a smaller size than the previous one. ([#303](https://github.com/ksnip/kImageAnnotator/issues/303))